Data.Time.CalendarTime

Calendar Time

data CalendarTime

A representation of calendar time separated into year, month, day, and so on.

Constructors

CalendarTime

Fields

calendarSecond :: Int

calendarMinute :: Int

calendarHour :: Int

calendarDay :: Int

calendarMonth :: Month

calendarYear :: Integer

calendarWeekDay :: WeekDay

calendarYearDay :: Int

calendarTimeZone :: TimeZone

toDay :: CalendarTime -> Maybe Day

Convert to a Day

withDay :: CalendarTime -> Day -> CalendarTime

Change y-m-d in CalendarTime

toTimeOfDay :: CalendarTime -> Maybe TimeOfDay

Convert to a TimeOfDay

daysInYear :: CalendarTimeConvertible a => a -> Int

lastDayOfMonth :: CalendarTimeConvertible a => a -> Int

weekNumber :: CalendarTimeConvertible a => StartOfWeek -> a -> Maybe Int

Calendar Time Convertible

class CalendarTimeConvertible t where

The class of types which can be converted to a CalendarTime

Methods

toCalendarTime :: t -> CalendarTime

Convert to a CalendarTime

fromCalendarTime :: CalendarTime -> Maybe t

Convert from a CalendarTime