Data.Time.Recurrence

The WeekDay type

data WeekDay

Symbolic week days.

Note: The first Day of the Week is Monday TODO: Move this to a more general library

Constructors

Monday

Tuesday

Wednesday

Thursday

Friday

Saturday

Sunday

The Month type

data Month

Symbolic months.

TODO: Move this to a more general library

Constructors

January

February

March

April

May

June

July

August

September

October

November

December

The Recurrence type

toUTCTime :: Recurrence -> UTCTime

Frequency Constrcutors for the Recuurence type

secondly :: Recurrence

Base Recurrence constructors. One for each of the Frequency types.

minutely :: Recurrence

hourly :: Recurrence

daily :: Recurrence

weekly :: Recurrence

monthly :: Recurrence

yearly :: Recurrence

Generates list of Recurrence type

recur :: Recurrence -> [Recurrence]

Generate a infinite list of recurrences

Recurrence type rule effects

withRules :: [Recurrence -> [Recurrence]] -> [Recurrence] -> [Recurrence]

Takes a list of rule parts and apply them over a list of Recurrence to generate a new list of Recurrence

byMonth :: [Month] -> Recurrence -> [Recurrence]

byWeekNumber :: [Int] -> Recurrence -> [Recurrence]

byYearDay :: [Int] -> Recurrence -> [Recurrence]