Data.Time.Recurrence.Schedule

Documentation

type Schedule a = Reader (InitialMoment a) [a]

recur :: a -> a

starting :: (Ord a, Moment a) => InitialMoment a -> a -> ([a] -> Schedule a) -> [a]

starting is an infinite list of Moments, where no Moment occurrs before the InitialMoment. The list is further refined by the passed in function.

begin :: (Ord a, Moment a) => InitialMoment a -> a -> [a]

enumMonths :: (CalendarTimeConvertible a, Moment a) => [Month] -> [a] -> Schedule a

enumDays :: (CalendarTimeConvertible a, Moment a) => [Int] -> [a] -> Schedule a

enumWeekDaysInWeek :: (CalendarTimeConvertible a, Moment a) => [WeekDay] -> [a] -> Schedule a

enumWeekDaysInMonth :: (CalendarTimeConvertible a, Moment a) => [WeekDay] -> [a] -> Schedule a

enumYearDays :: (CalendarTimeConvertible a, Moment a) => [Int] -> [a] -> Schedule a

enumHours :: (CalendarTimeConvertible a, Moment a) => [Int] -> [a] -> Schedule a

enumMinutes :: (CalendarTimeConvertible a, Moment a) => [Int] -> [a] -> Schedule a

enumSeconds :: (CalendarTimeConvertible a, Moment a) => [Int] -> [a] -> Schedule a

nthMonth :: CalendarTimeConvertible a => [Int] -> [a] -> Schedule a

nthDay :: CalendarTimeConvertible a => [Int] -> [a] -> Schedule a

nthWeekDay :: CalendarTimeConvertible a => [Int] -> [a] -> Schedule a

nthWeekDayOfWeek :: CalendarTimeConvertible a => [Int] -> [a] -> Schedule a

nthWeekDayOfMonth :: CalendarTimeConvertible a => [Int] -> [a] -> Schedule a

nthYearDay :: CalendarTimeConvertible a => [Int] -> [a] -> Schedule a

nthHour :: CalendarTimeConvertible a => [Int] -> [a] -> Schedule a

nthMinute :: CalendarTimeConvertible a => [Int] -> [a] -> Schedule a

nthSecond :: CalendarTimeConvertible a => [Int] -> [a] -> Schedule a

filterMonths :: CalendarTimeConvertible a => [Month] -> [a] -> Schedule a

filterDays :: CalendarTimeConvertible a => [Int] -> [a] -> Schedule a

filterWeekDays :: CalendarTimeConvertible a => [WeekDay] -> [a] -> Schedule a

filterYearDays :: CalendarTimeConvertible a => [Int] -> [a] -> Schedule a

filterHours :: CalendarTimeConvertible a => [Int] -> [a] -> Schedule a

filterMinutes :: CalendarTimeConvertible a => [Int] -> [a] -> Schedule a

filterSeconds :: CalendarTimeConvertible a => [Int] -> [a] -> Schedule a