Data.Time.Recurrence

The WeekDay type

data WeekDay

Symbolic week days.

Note: The first Day of the Week is Monday TODO: Move this to a more general library

Constructors

Monday

Tuesday

Wednesday

Thursday

Friday

Saturday

Sunday

The Month type

data Month

Symbolic months.

TODO: Move this to a more general library

Constructors

January

February

March

April

May

June

July

August

September

October

November

December

The Moment type class

class Moment a where

The Moment class is for representing a instance in time.

Instances of Moment can be derived for any user-defined datatype for which can satisfy the minimal complete definition.

Minimal complete definition: epoch, toDateTime, fromDateTime, scaleTime, scaleMonth, scaleYear, alterWeekNumber, alterYearDay

Methods

epoch :: a

Provide a default moment.

toDateTime :: a -> DateTime

Convert a Moment into a DateTime

fromDateTime :: DateTime -> Maybe a

Convert a DateTime into a Moment

scaleTime :: a -> Integer -> a

Produce a new Moment offset by a given number of seconds.

scaleMonth :: a -> Integer -> a

Produce a new Moment offset by a given number of months.

scaleYear :: a -> Integer -> a

Produce a new Moment offset by a given number of years.

alterWeekNumber :: StartOfWeek -> a -> Int -> Maybe a

Possibly produce a new Moment shifted to a different week of the year.

alterYearDay :: a -> Int -> Maybe a

Possibly produce a new Moment shifted to a different day of the year.

alterSecond :: a -> Int -> Maybe a

The 'alter*' methods can potentially produce invalid dates.

For each user-defined Moment instance the definitions of toDateTime, fromDateTime, alterWeekNumber and alterYearDay will determine if an altered Moment that lands on an invalid date in the given calendar will be reduced to Nothing

Possibly produce a new Moment shifted to a different second of the day.

alterMinute :: a -> Int -> Maybe a

Possibly produce a new Moment shifted to a different minute of the day.

alterHour :: a -> Int -> Maybe a

Possibly produce a new Moment shifted to a different hour of the day.

alterDay :: a -> Int -> Maybe a

Possibly produce a new Moment shifted to a different day of the month.

alterMonth :: a -> Month -> Maybe a

Possibly produce a new Moment shifted to a different month of the year.

alterYear :: a -> Integer -> Maybe a

Possibly produce a new Moment shifted to a different year.

next :: Interval -> Frequency -> a -> a

Produce a new Moment in the future ocurring at (interval * freq)

prev :: Interval -> Frequency -> a -> a

Produce a new Moment in the past ocurring at (-interval * freq)

The InitialMoment type

data InitialMoment a

The InitialMoment datatype

Constructors

InitialMoment

Fields

frequency :: Frequency

interval :: Interval

startOfWeek :: StartOfWeek

moment :: a

toInterval :: Integer -> Interval

toStartOfWeek :: WeekDay -> StartOfWeek

secondly :: Moment a => InitialMoment a

Default initial moments

minutely :: Moment a => InitialMoment a

hourly :: Moment a => InitialMoment a

daily :: Moment a => InitialMoment a

weekly :: Moment a => InitialMoment a

monthly :: Moment a => InitialMoment a

yearly :: Moment a => InitialMoment a

secondlyUTC :: InitialMoment UTCTime

InitialMoment defaults for UTCTime

minutelyUTC :: InitialMoment UTCTime

hourlyUTC :: InitialMoment UTCTime

dailyUTC :: InitialMoment UTCTime

weeklyUTC :: InitialMoment UTCTime

monthlyUTC :: InitialMoment UTCTime

yearlyUTC :: InitialMoment UTCTime

Recurrence combinators

enumYear :: (Moment a, Ord a) => a -> RecurringSchedule a

enumYear produces all days in the year starting with m

enumMonth :: (Moment a, Ord a) => a -> RecurringSchedule a

enumMonth produces all days in the current month starting with m

enumWeek :: (Moment a, Ord a) => a -> RecurringSchedule a

enumWeek produces all days in the current week starting with m

restrict :: Moment a => (a -> Bool) -> Schedule a -> RecurringSchedule a

restrict, applied to a predicate and a Schedule, returns a Schedule of those moments that statisfy the predicate.

bySeconds :: Moment a => [Int] -> a -> Bool

byMinutes :: Moment a => [Int] -> a -> Bool

byHours :: Moment a => [Int] -> a -> Bool

byWeekDays :: Moment a => [WeekDay] -> a -> Bool

byMonthDays :: Moment a => [Int] -> a -> Bool

byMonths :: Moment a => [Month] -> a -> Bool

byYearDays :: Moment a => [Int] -> a -> Bool

expand :: Moment a => (a -> Reader (InitialMoment a) [a]) -> Schedule a -> RecurringSchedule a

expand, takes an expansion function and a Schedule, and maps the expansion function over the moments. Each moment is then replaced with its expansions.

onWeekNumbers :: Moment a => [Int] -> a -> Reader (InitialMoment a) [a]

onMonths :: Moment a => [Month] -> a -> Reader (InitialMoment a) [a]

onMonthDays :: Moment a => [Int] -> a -> Reader (InitialMoment a) [a]

onYearDays :: Moment a => [Int] -> a -> Reader (InitialMoment a) [a]

onEachWeek :: (Moment a, Ord a) => a -> Reader (InitialMoment a) [a]

onEachMonth :: (Moment a, Ord a) => a -> Reader (InitialMoment a) [a]

onEachYear :: (Moment a, Ord a) => a -> Reader (InitialMoment a) [a]

repeatSchedule :: Moment a => InitialMoment a -> (Schedule a -> RecurringSchedule a) -> [a]

repeatSchedule runs a schedule with a given init and rule r

repeatSchedule' :: Moment a => InitialMoment a -> [a]

repeatSchedule' is like repeatSchedule but it takes no rules