Data.Time.Recurrence

Documentation

data CalendarTime

A representation of calendar time separated into year, month, day, and so on.

Constructors

CalendarTime

Fields

calendarSecond :: Int

calendarMinute :: Int

calendarHour :: Int

calendarDay :: Int

calendarMonth :: Month

calendarYear :: Integer

calendarWeekDay :: WeekDay

calendarYearDay :: Int

calendarTimeZone :: TimeZone

class Moment a

The Moment class is for representing a instance in time.

Instances of Moment can be derived for any user-defined datatype for which can satisfy the minimal complete definition.

Minimal complete definition: epoch, addSeconds, addMonths, addYears