Data.Time.Recurrence

The DateTime type

data DateTime

DateTime data type This is a componentized version of a time value simmilar to a 'struct tm'

Constructors

DateTime

Fields

dtSecond :: Int

dtMinute :: Int

dtHour :: Int

dtDay :: Int

dtMonth :: Month

dtYear :: Integer

dtWeekDay :: WeekDay

dtYearDay :: Int

dtTimeZone :: TimeZone

The WeekDay type

data WeekDay

Symbolic week days.

Note: The first Day of the Week is Monday TODO: Move this to a more general library

Constructors

Monday

Tuesday

Wednesday

Thursday

Friday

Saturday

Sunday

The Month type

data Month

Symbolic months.

TODO: Move this to a more general library

Constructors

January

February

March

April

May

June

July

August

September

October

November

December

The Moment type class

class Ord a => Moment a where

Moment type class

Methods

epoch :: a

Minimum Definition

toDateTime :: a -> DateTime

fromDateTime :: DateTime -> Maybe a

scaleTime :: a -> Integer -> a

scaleMonth :: a -> Integer -> a

scaleYear :: a -> Integer -> a

alterWeekNumber :: StartOfWeek -> a -> Int -> Maybe a

alterYearDay :: a -> Int -> Maybe a

Other Methods

alterSecond :: a -> Int -> Maybe a

alterMinute :: a -> Int -> Maybe a

alterHour :: a -> Int -> Maybe a

alterDay :: a -> Int -> Maybe a

alterMonth :: a -> Month -> Maybe a

alterYear :: a -> Integer -> Maybe a

next :: Interval -> Frequency -> a -> a

prev :: Interval -> Frequency -> a -> a

The RecurrenceParameters type

data RecurrenceParameters a

The RecurrenceParameters type

Constructors

RecurrenceParameters

Fields

interval :: Interval

frequency :: Frequency

startOfWeek :: StartOfWeek

startDate :: a

secondly

Arguments

:: Moment a

=> RecurrenceParameters a
Recur every second

Some common recurrence parameters

minutely

Arguments

:: Moment a

=> RecurrenceParameters a
Recur every minute

hourly

Arguments

:: Moment a

=> RecurrenceParameters a
Recur every hour

daily

Arguments

:: Moment a

=> RecurrenceParameters a
Recur every day

weekly

Arguments

:: Moment a

=> RecurrenceParameters a
Recur every week

monthly

Arguments

:: Moment a

=> RecurrenceParameters a
Recur every month

yearly

Arguments

:: Moment a

=> RecurrenceParameters a
Recur every year

create types in RecurrenceParameters

toInterval :: Integer -> Interval

toStartOfWeek :: WeekDay -> StartOfWeek

Generate list of recurring Moment

recur

Arguments

:: Moment a

=> [RecurrenceParameters a -> a -> Recurrence a]
Sub Rules on Moments

-> RecurrenceParameters a
Parameters of the recurrence

-> Recurrence a
Resulting Recurrence

count :: Moment a => Int -> Recurrence a -> Recurrence a

Limit recurrence results to a count

until :: Moment a => a -> Recurrence a -> Recurrence a

Limit recurrence results to a max date

The Recurrence type

data Recurrence a

The Recurrence type a Recurrence is on or more Moments

Recurrence type rule effects

byMonth :: Moment a => [Month] -> RecurrenceParameters a -> a -> Recurrence a

byWeekNumber :: Moment a => [Int] -> RecurrenceParameters a -> a -> Recurrence a

byYearDay :: Moment a => [Int] -> RecurrenceParameters a -> a -> Recurrence a

byMonthDay :: Moment a => [Int] -> RecurrenceParameters a -> a -> Recurrence a

byDay :: Moment a => [WeekDay] -> RecurrenceParameters a -> a -> Recurrence a